requirements.txt
fastapi
uvicorn
google-cloud-firestore
python-dotenv
newspaper3k
lxml[html_clean]